fix(welcome): address coderabbit follow-up on PR #894
- Move webview_accounts detector + tests into a sibling `ops.rs` so
  `mod.rs` stays the export-focused file the repo conventions ask for.
- Drop `path.display()` from the two debug logs in the cookies-DB
  open/exists branches; log the env key instead so usernames in
  /Users/<name>/... paths can't end up in trace output.
- Restore lowercase `linkedin` in the welcome prompt — the JSON key
  exported by `webview_logins` is lowercase, capitalising it would
  desync the agent's pattern match from the actual snapshot field.
